// packages/destination-actions/src/destinations/braze/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Settings {
  /**
   * Created under Developer Console in the Braze Dashboard.
   */
  api_key: string
  /**
   * The app identifier used to reference specific Apps in requests made to the Braze API. Created under Developer Console in the Braze Dashboard.
   */
  app_id: string
  /**
   * Your Braze REST endpoint. [See more details](https://www.braze.com/docs/api/basics/#endpoints)
   */
  endpoint: string
}