// packages/destination-actions/src/destinations/braze/createAlias2/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Payload {
  /**
   * The external ID of the user to create an alias for.
   */
  external_id?: string | null
  /**
   * The alias identifier
   */
  alias_name: string
  /**
   * A label indicating the type of alias
   */
  alias_label: string
}